separate ICapacity
